refactor the extension into separate hooks
